% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Sensabled Shiny Application}
\usage{
run_app(...)
}
\arguments{
\item{...}{Arguments passed to shinyAppDir}
}
\value{
A Shiny app object
}
\description{
Run the Sensabled Shiny Application
}
